Remove auto-scoping of errors and separate derivable in a submodule
